+++ R/as_ggsurvplot_df.r
@@ -20,12 +20,47 @@ as_ggsurvplot_df <- function(adjsurv) {
     df$lower <- adjsurv$adj$ci_lower
   }
 
-  if (adjsurv$method=="iptw_km") {
-    df$n.risk <- adjsurv$n_at_risk$n_at_risk
-    df$n.event <- adjsurv$n_at_risk$n_events
-  } else if (adjsurv$method=="km") {
-    df$n.risk <- adjsurv$survfit_object$n.risk
-    df$n.event <- adjsurv$survfit_object$n.event
+  if (!is.null(adjsurv$mids_analyses)) {
+
+    # set correct weights if specified
+    if (!is.null(adjsurv$weights) && is.null(adjsurv$mids_analyses)) {
+      weights <- adjsurv$weights
+    } else if (!is.null(adjsurv$mids_analyses) &&
+               !is.null(adjsurv$mids_analyses[[1]]$weights)) {
+      weights <- lapply(adjsurv$mids_analyses, FUN=function(d){d$weights})
+    } else {
+      weights <- NULL
+    }
+
+    # calculate pooled risk table
+    n.risk <- get_risk_table(times=df$time,
+                             data=adjsurv$data,
+                             ev_time=adjsurv$ev_time,
+                             variable=adjsurv$variable,
+                             event=adjsurv$event,
+                             type="n_at_risk",
+                             weights=weights,
+                             digits=Inf)
+    colnames(n.risk) <- c("time", "strata", "n.risk")
+
+    n.risk$n.event <- get_risk_table(times=df$time,
+                                     data=adjsurv$data,
+                                     ev_time=adjsurv$ev_time,
+                                     variable=adjsurv$variable,
+                                     event=adjsurv$event,
+                                     type="n_events",
+                                     weights=weights,
+                                     digits=Inf)$est
+    df <- merge(df, n.risk, by=c("time", "strata"))
+
+  } else {
+    if (adjsurv$method=="iptw_km") {
+      df$n.risk <- adjsurv$n_at_risk$n_at_risk
+      df$n.event <- adjsurv$n_at_risk$n_events
+    } else if (adjsurv$method=="km") {
+      df$n.risk <- adjsurv$survfit_object$n.risk
+      df$n.event <- adjsurv$survfit_object$n.event
+    }
   }
 
   return(df)