+++ NEWS.md
@@ -1,4 +1,11 @@
 
+# adjustedCurves 0.11.4 (developmental version)
+
+Bug Fixes
+
+* Fixed a bug that occurred when using `plot.adjustedsurv()` with the `times` argument in case a dataset with multiple imputation was supplied.
+* Numbers at risk and number of events are now included in the output of `as_ggsurvplot_df()`, even when `method = "km"` was used. Additionally, with multiple imputation, the correctly pooled values are now included, instead of nothing.
+
 # adjustedCurves 0.11.3
 
 New features